requirements.txt
eth-brownie==1.14.6
PyQt5
ganache-cli